pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rgbd-depth"
version = "1.0.2"
description = "Optimized RGB-D depth refinement using Vision Transformers with CUDA, MPS, and CPU support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "ByteDance Research"},
    {name = "Aedelon", email = "delanoe.pirard.pro@gmail.com"}
]
maintainers = [
    {name = "Aedelon", email = "delanoe.pirard.pro@gmail.com"}
]
keywords = [
    "depth-estimation",
    "rgbd",
    "vision-transformer",
    "pytorch",
    "computer-vision",
    "robotics",
    "sim-to-real",
    "xformers",
    "mixed-precision"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing"
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.5.0",
    "numpy>=1.20.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
# CUDA optimizations (xFormers for ~8% speedup)
xformers = [
    "xformers>=0.0.20; platform_system != 'Darwin'",
]

# Gradio demo
demo = [
    "gradio>=4.0.0",
    "matplotlib>=3.5.0",
]

# Development tools
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
]

# Model downloads from HuggingFace
download = [
    "huggingface-hub>=0.16.0",
]

# Everything
all = [
    "xformers>=0.0.20; platform_system != 'Darwin'",
    "gradio>=4.0.0",
    "matplotlib>=3.5.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "huggingface-hub>=0.16.0",
]

[project.urls]
Homepage = "https://github.com/Aedelon/camera-depth-models"
Repository = "https://github.com/Aedelon/camera-depth-models"
Documentation = "https://github.com/Aedelon/camera-depth-models/blob/main/README.md"
"Bug Tracker" = "https://github.com/Aedelon/camera-depth-models/issues"
"Original Research" = "https://github.com/bytedance/camera-depth-models"
"Paper" = "https://manipulation-as-in-simulation.github.io/"

[project.scripts]
rgbd-depth = "infer:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rgbddepth*"]
exclude = ["tests*", "docs*", "example_data*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
